+++ test/unit/fuzz-tools.test.js
@@ -124,6 +124,7 @@ import * as psychV123 from '../../lib/psych-v123.js';
 import * as hepV124 from '../../lib/hep-v124.js';
 import * as hepV125 from '../../lib/hep-v125.js';
 import * as giV126 from '../../lib/gi-v126.js';
+import * as nephroV127 from '../../lib/nephro-v127.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -192,6 +193,7 @@ const MODULES = {
   'hep-v124.js': hepV124,
   'hep-v125.js': hepV125,
   'gi-v126.js': giV126,
+  'nephro-v127.js': nephroV127,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];